FIX reduce heap memory pressure on a driver
The main idea is to read CompressionInfo files directly from mappers.
In such way we reduce an amount of memory in which we store input splits.
This should help to the driver program (where splitting is happening)
stay cool on big data sets.

Would love to perform splitting as one more step in the MR job,
but let's leave this for future.